First push

+++ ui/OrdercarUi.py
@@ -23,7 +23,7 @@ class OrdercarUi:
     @staticmethod
     def print_current_orders(orders):
         if len(orders) == 0:
-            print("No orders\n")
+            print("\nNo orders\n")
         else:
             print("{:^6}|{:^12}|{:^20}|{:^17}|{:^21}|{:^21}|{:^12}|{:^11}|{:^13}|{:^6}".format
                   ("ID", "Kt", "Name", "Car-license", "From date", "To date", "Price", "Insurance", "Total price", "Days"))
@@ -35,8 +35,8 @@ class OrdercarUi:
 
     @staticmethod
     def print_completed_orders(completed_orders):
-        if len(completed_orders) == 0:
-            print("No orders")
+        if completed_orders == "No orders":
+            print("\nNo orders\n")
         else:
             print("{:^6}|{:^12}|{:^20}|{:^17}|{:^21}|{:^21}|{:^20}|{:^21}|{:^11}|{:^13}|{:^6}".format
                   ("ID", "Kt", "Name", "License", "From date", "To date", "Price", "Payment method", "Insurance",
@@ -119,76 +119,80 @@ class OrdercarUi:
 
         approved = False
         while not approved:
-            from_date = self.__car_service.user_date("\tEnter start date for rent (dd/mm/yy): ")
-            to_date = self.__car_service.user_date("\tEnter end date for rent (dd/mm/yy): ")
-            for x in self.__car_service.get_car_class():
-                print(str(x), end=' ')
-            print()
-            car_type = input("\tEnter type of car (q to quit): ").translate(remove_punct_map)
-            if car_type.upper() == "Q":
-                break
-            print("Available cars\n")
-
-            available_cars_type = self.__car_service.get_available_date_type(car_type, from_date, to_date)
-
-            if len(available_cars_type) == 0:
-                i = input("No cars available,(Press q to quit, enter to select another date)")
-                if i == "q":
+            try:
+                from_date = self.__car_service.user_date("\tEnter start date for rent (dd/mm/yy): ")
+                to_date = self.__car_service.user_date("\tEnter end date for rent (dd/mm/yy): ")
+                for x in self.__car_service.get_car_class():
+                    print(str(x), end=' ')
+                print()
+                car_type = input("\tEnter type of car (q to quit): ").translate(remove_punct_map)
+                if car_type.upper() == "Q":
                     break
-            else:
-                while not approved:
-                    self.__car_ui.print_cars(available_cars_type)
-                    try:
-                        c_id = input("\nSelect car by Id (q to quit): ").upper()
-                        if c_id == "Q":
-                            approved = True
-                            break
-                        c_id = int(c_id)
-                        self.__car_ui.print_cars([available_cars_type[c_id - 1]])
+                print("Available cars")
 
-                        chosen_car_plate = available_cars_type[c_id - 1]["License"]
-                        price_of_order = int(available_cars_type[c_id - 1]["Price"])
+                available_cars_type = self.__car_service.get_available_date_type(car_type, from_date, to_date)
 
-                        # Calculate how long the order is in days
-                        from_date = datetime.datetime.date(from_date)
-                        to_date = datetime.datetime.date(to_date)
-                        delta = to_date - from_date
-                        days = delta.days   # how many days the rental is
+                if len(available_cars_type) == 0:
+                    i = input("No cars available,(Press q to quit, enter to select another date)")
+                    if i == "q":
+                        break
+                else:
+                    while not approved:
+                        self.__car_ui.print_cars(available_cars_type)
+                        try:
+                            c_id = input("\nSelect car by Id (q to quit): ").upper()
+                            if c_id == "Q":
+                                approved = True
+                                break
+                            c_id = int(c_id)
+                            self.__car_ui.print_cars([available_cars_type[c_id - 1]])
 
-                        price_of_order_days = price_of_order * days     # Price for car multiplied with days
-                        print("Price of order: {} ISK".format(price_of_order_days))
-                        insurance = input("Would you like extra insurance for {} ISK per day? Y/N: ".format(
-                            int(price_of_order) * 0.75)).upper()    # Insurance (Yes or No)
-                        price_of_order_days_insurance = price_of_order_days
-                        if insurance == 'Y':
-                            price_of_order_days_insurance = price_of_order_days * 1.75  # Price of order with extra insurance
-                            print("Price of order: {} ISK".format(price_of_order_days_insurance))
-                            deposit = price_of_order_days_insurance * 0.10
-                            print("Your deposit of the order is {} ISK".format(deposit))
-                        else:
-                            deposit = price_of_order_days * 0.10
-                            print("Your deposit of the order is {} ISK".format(deposit))
+                            chosen_car_plate = available_cars_type[c_id - 1]["License"]
+                            price_of_order = int(available_cars_type[c_id - 1]["Price"])
 
-                        book = input("Order car? Y/N: ").upper()
-                        if book == 'Y':
-                            if customer:
-                                name = customer["Name"]
+                            # Calculate how long the order is in days
+                            from_date = datetime.datetime.date(from_date)
+                            to_date = datetime.datetime.date(to_date)
+                            delta = to_date - from_date
+                            days = delta.days   # how many days the rental is
 
-                            new_order = Order(kt, name, chosen_car_plate, from_date, to_date, price_of_order_days,
-                                              insurance, price_of_order_days_insurance, days)
-                            self.__order_service.add_order(new_order, False)
-                            print("\nOrder successful!\n")
-                            approved = True
-                        else:
-                            continue
-                    except IndexError:
-                        print("ID not available")
+                            price_of_order_days = price_of_order * days     # Price for car multiplied with days
+                            print("Price of order: {} ISK".format(price_of_order_days))
+                            insurance = input("Would you like extra insurance for {} ISK per day? Y/N: ".format(
+                                int(price_of_order) * 0.75)).upper()    # Insurance (Yes or No)
+                            price_of_order_days_insurance = price_of_order_days
+                            if insurance == 'Y':
+                                price_of_order_days_insurance = price_of_order_days * 1.75  # Price of order with extra insurance
+                                print("Price of order: {} ISK".format(price_of_order_days_insurance))
+                                deposit = price_of_order_days_insurance * 0.10
+                                print("Your deposit of the order is {} ISK".format(deposit))
+                            else:
+                                deposit = price_of_order_days * 0.10
+                                print("Your deposit of the order is {} ISK".format(deposit))
+
+                            book = input("Order car? Y/N: ").upper()
+                            if book == 'Y':
+                                if customer:
+                                    name = customer["Name"]
+
+                                new_order = Order(kt, name, chosen_car_plate, from_date, to_date, price_of_order_days,
+                                                  insurance, price_of_order_days_insurance, days)
+                                self.__order_service.add_order(new_order, False)
+                                print("\nOrder successful!\n")
+                                approved = True
+                            else:
+                                continue
+                        except IndexError:
+                            print("ID not available")
+            except Exception:
+                print("\nNo cars available\n")
+                break
 
     def return_car(self):
         try:
             orders = self.__order_service.get_orders()
             if len(orders) == 0:
-                print("\nNo orders")
+                print("\nNo orders\n")
             else:
                 self.print_current_orders(orders)
                 o_id = int(input("Select order by Id: "))
@@ -205,7 +209,7 @@ class OrdercarUi:
         try:
             orders = self.__order_service.get_orders()
             if len(orders) == 0:
-                print("\nNo orders")
+                print("\nNo orders\n")
             else:
                 self.print_current_orders(orders)
                 o_id = int(input("Select order by Id (q to quit): "))
@@ -220,34 +224,37 @@ class OrdercarUi:
                 elif choice == 'n':
                     print("Revoke canceled")
         except Exception:
-            print("Canceled")
+            print("Revoke Canceled")
 
     def edit_current_order(self):
         orders = self.__order_service.get_orders()
-        self.print_current_orders(orders)
-        o_id = int(input("Select order by Id: "))
-        order = self.__order_service.get_order_by_id(o_id)
-        edited_order = Order(order["Kt"], order["Name"], order["License"], order["From date"], order["To date"],
-                             order["Price"], order["Insurance"], order["Total price"], order["Days"])
-        a_choice = ''
-        while a_choice != 'q':
-            a_choice = input(
-                "1. Edit name\n2. License\n3. From date\n4. To date\n5. Price\nPress q to go back").lower()
-            if a_choice == '1':
-                edited_order.set_renter(input("Enter new name: ").translate(remove_punct_map))
-            elif a_choice == '2':
-                edited_order.set_car(input("Enter new license: ").translate(remove_punct_map))
-            elif a_choice == '3':
-                edited_order.set_from_date(
-                    datetime.datetime.strftime(self.__car_service.user_date("Enter new from date: "), "%d/%m/%y"))
-            elif a_choice == '4':
-                edited_order.set_to_date(
-                    datetime.datetime.strftime(self.__car_service.user_date("Enter new to date: "), "%d/%m/%y"))
-            elif a_choice == '5':
-                edited_order.set_price(input("Enter new price: ").translate(remove_punct_map))
-        print(edited_order)
-        self.__order_service.remove_order(o_id)
-        self.__order_service.add_order(edited_order, True)
+        if orders != '':
+            self.print_current_orders(orders)
+            o_id = int(input("Select order by Id: "))
+            order = self.__order_service.get_order_by_id(o_id)
+            edited_order = Order(order["Kt"], order["Name"], order["License"], order["From date"], order["To date"],
+                                 order["Price"], order["Insurance"], order["Total price"], order["Days"])
+            a_choice = ''
+            while a_choice != 'q':
+                a_choice = input(
+                    "1. Edit name\n2. License\n3. From date\n4. To date\n5. Price\nPress q to go back").lower()
+                if a_choice == '1':
+                    edited_order.set_renter(input("Enter new name: ").translate(remove_punct_map))
+                elif a_choice == '2':
+                    edited_order.set_car(input("Enter new license: ").translate(remove_punct_map))
+                elif a_choice == '3':
+                    edited_order.set_from_date(
+                        datetime.datetime.strftime(self.__car_service.user_date("Enter new from date: "), "%d/%m/%y"))
+                elif a_choice == '4':
+                    edited_order.set_to_date(
+                        datetime.datetime.strftime(self.__car_service.user_date("Enter new to date: "), "%d/%m/%y"))
+                elif a_choice == '5':
+                    edited_order.set_price(input("Enter new price: ").translate(remove_punct_map))
+            print(edited_order)
+            self.__order_service.remove_order(o_id)
+            self.__order_service.add_order(edited_order, True)
+        else:
+            print("\nNo orders to edit\n")
 
     def edit_completed_order(self):
         orders = self.__order_service.get_completed_orders()
@@ -293,7 +300,7 @@ class OrdercarUi:
             print("5. Revoke order")
             print("6. Edit order")
             print("7. List order history of car")
-            print("Press q to quit")
+            print("Press q to go back")
 
             action = input()
             if action == '1':
@@ -308,15 +315,17 @@ class OrdercarUi:
                 input("Press enter to continue")
 
             elif action == '4':
-                completed_orders = self.__order_service.get_completed_orders()
-                self.print_completed_orders(completed_orders)
-                o_id = input("Select the order you want to view (q to quit): ")
-                if o_id == "q":
-                    break
-                os.system('cls')
-                order = self.__order_service.get_completed_order_id(int(o_id))
-                self.print_receipt(order)
-
+                try:
+                    completed_orders = self.__order_service.get_completed_orders()
+                    if self.print_completed_orders(completed_orders) == "No orders":
+                        o_id = ''
+                        while o_id != 'q':
+                            o_id = input("Select the order you want to view (q to quit): ")
+                            os.system('cls')
+                            order = self.__order_service.get_completed_order_id(int(o_id))
+                            self.print_receipt(order)
+                except Exception:
+                    print("Something went wrong")
                 input("Press enter to continue")
 
             elif action == '5':
@@ -331,10 +340,14 @@ class OrdercarUi:
                         self.edit_current_order()
                     elif e_action == '2':
                         self.edit_completed_order()
+                input("Press enter to continue")
 
             elif action == "7":
-                license = input("Enter car license plate (q to quit): ").upper()
-                if license != "Q":
-                    orders = self.__order_service.get_available_orders(license)
-                    self.print_completed_orders(orders)
+                if self.__order_service.get_orders() != '':
+                    license = input("Enter car license plate (q to quit): ").upper()
+                    if license != "Q":
+                        orders = self.__order_service.get_available_orders(license)
+                        self.print_completed_orders(orders)
+                else:
+                    print("\nNo orders\n")
                 input("Press enter to continue")
